refactor: 매직넘버 상수로 변경
