+++ src/main/kotlin/blackjack/domain/GameResult.kt
@@ -1,26 +1,5 @@
 package blackjack.domain
 
-class DealerResult() {
-    var win: Int = 0
-        private set
-    var lose: Int = 0
-        private set
-    var draw: Int = 0
-        private set
-
-    fun addWin() {
-        win++
-    }
-
-    fun addLose() {
-        lose++
-    }
-
-    fun addDraw() {
-        draw++
-    }
-}
-
 class GameResult(private val dealer: Dealer, val players: List<Player>) {
     val dealerResult = DealerResult()
     val playerResultStatus = mutableMapOf<Player, GameResultStatus>()
